requirements.txt
django==5.1.2
pillow==11.0.0
psycopg==3.2.3
requests==2.32.3
python-dotenv==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1


